+++ src/api/scope/lib/delete.js
@@ -1,12 +1,16 @@
-// @flow
+/** @flow */
 import { loadScope } from '../../../scope';
 import { BitIds } from '../../../bit-id';
 import { PRE_REMOVE_REMOTE, POST_REMOVE_REMOTE } from '../../../constants';
 import HooksManager from '../../../hooks';
+import { RemovedObjects } from '../../../scope/removed-components';
 
 const HooksManagerInstance = HooksManager.getInstance();
 
-export default function remove({ path, ids, force }, headers: ?Object): Promise<string[]> {
+export default function remove(
+  { path, ids, force }: { path: string, ids: string[], force: boolean },
+  headers: ?Object
+): Promise<RemovedObjects> {
   const bitIds = BitIds.deserialize(ids);
   const args = { path, bitIds, force };
   HooksManagerInstance.triggerHook(PRE_REMOVE_REMOTE, args, headers);
@@ -15,7 +19,7 @@ export default function remove({ path, ids, force }, headers: ?Object): Promise<
       await HooksManagerInstance.triggerHook(
         POST_REMOVE_REMOTE,
         {
-          removedComponentsIds: res.bitIds,
+          removedComponentsIds: res.removedComponentIds,
           missingComponentsIds: res.missingComponents,
           dependentBitsIds: res.dependentBits,
           force,